# api/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/change_pwd": {
            "post": {
                "description": "管理员密码修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员密码修改",
                "operationId": "/admin/change_pwd",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminChangePwdInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "管理员登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "管理员登陆",
                "operationId": "/admin/login",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminLoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AdminLoginOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/app_add": {
            "post": {
                "description": "APP添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP添加",
                "operationId": "/app/app_add",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.APPAddHttpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/app_delete": {
            "get": {
                "description": "APP删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP删除",
                "operationId": "/app/app_delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APP ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/app_detail": {
            "get": {
                "description": "APP详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP详情",
                "operationId": "/app/app_detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APP ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/app_list": {
            "get": {
                "description": "APP列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP列表",
                "operationId": "/app/app_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_no",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.APPListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/app_stat": {
            "get": {
                "description": "APP统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP统计",
                "operationId": "/app/app_stat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "APP ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.StatisticsOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/app_update": {
            "post": {
                "description": "APP更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APP"
                ],
                "summary": "APP更新",
                "operationId": "/app/app_update",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.APPUpdateHttpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dao.App": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
...
        "dto.APPAddHttpInput": {
            "type": "object",
            "required": [
                "app_id",
                "name"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.APPListItemOutput": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "real_qpd": {
                    "type": "integer"
                },
                "real_qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.APPListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.APPListItemOutput"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.APPUpdateHttpInput": {
            "type": "object",
            "required": [
                "id",
                "name",
                "secret"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.AdminChangePwdInput": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "dto.AdminLoginInput": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "description": "管理员用户名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "dto.AdminLoginOutput": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "token",
                    "type": "string",
                    "example": "token"
                }
            }
        },
...
        "dto.StatisticsOutput": {
            "type": "object",
            "required": [
                "today",
                "yesterday"
            ],
            "properties": {
                "today": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "pkg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {},
                "trace_id": {}
            }
        }
    }
}